requirements.txt
amqp==2.6.1
asgiref==3.3.0
bcrypt==3.2.0
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
croniter==0.3.36
cryptography==3.3.1
Django==3.1.3
django-anymail==8.1
django-appconf==1.0.4
django-cryptography==1.0
django-debug-toolbar==3.1.1
django-extensions==3.0.9
django-redis-cache==3.0.0
djangorestframework==3.12.2
flower==0.9.7
gunicorn==20.0.4
humanize==3.2.0
idna==2.10
kombu==4.6.11
Markdown==3.3.3
markdown-urlize==0.2.0
natsort==7.0.1
paramiko==2.7.2
Pillow==8.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.4
redis==3.5.3
requests==2.25.0
sentry-sdk==0.19.5
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
tornado==6.1
urllib3==1.25.11
vine==1.3.0
watchgod==0.6
wcwidth==0.2.5
whitenoise==5.2.0
